% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_network_build.R
\name{assign_size_class}
\alias{assign_size_class}
\title{Assign size classes, for all species}
\usage{
assign_size_class(data, species, var, classes)
}
\arguments{
\item{data}{a data.frame containing species and size variable.}

\item{species}{species variable name in the dataset.}

\item{var}{variable characterizing the size.}

\item{classes}{data.frame created by the compute_classes function.}
}
\value{
data.frame containing class_id for each indivual.
}
\description{
Assign size classes, for all species
}
