% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaweb_build.R
\name{build_metaweb}
\alias{build_metaweb}
\title{Metaweb building}
\usage{
build_metaweb(data, species, size, pred_win, fish_diet_shift, low_bound,
  upper_bound, fish, resource_diet_shift, class_method = "quantile",
  nb_class = 9, pred_win_method = "midpoint", na.rm = FALSE,
  replace_min_by_one = FALSE, fish_resource_method = "overlap")
}
\arguments{
\item{data}{a dataframe containing species and size variable}

\item{species}{}

\item{size}{data.frame containing species and length.}

\item{pred_win}{data.frame containing species and the predation window
parameters (\alpha and \beta).}

\item{class_method}{character percentile or quantile. Default to percentile.}

\item{nb_class}{integer number of size class to create. Default to 9.}

\item{pred_win_method}{character midpoint, overlap or no_overlap. Default to midpoint.}

\item{na.rm}{logical. Should the NAs be removed ?}

\item{replace_min_by_one}{logical. In the first class, replace low bound by
1.}

\item{fish_resource_method}{character Either "overlap", "willem" or "midpoint"}

\item{diet_shift}{data.frame containing species and the predation window.}
}
\value{
matrix
}
\description{
Metaweb building
}
\details{
class_method = "quantile" use the `quantile` function to create size
 class in nb_class . When class_method = "percentile", the range of the size distribution
 is divided by nb_class. The difference is that `quantile` start the first
 class at the minimum value of the size distribution while percentile begins
 at 0. It therefore gives slightly different results.
 When the predation window definition is defined as overlap or midpoint, there
 will be possibility of overlap between size class of a predator and a prey.
 It means that for trophic link established between a prey and a predator, the
 smallest predator of a given size class is smaller than the biggest prey
 fish. When pred_win_method is set to "no_overlap", the upper limit of the predation window of
 a predator size class is defined by the smallest predator of the size class,
 so there is no overlap between predator and prey. With midpoint option, the
 existence of a trophic link is defined by the middle of the predator size
 class.
}
