% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaweb_build.R
\name{compute_classes}
\alias{compute_classes}
\title{Compute classes}
\usage{
compute_classes(size, group_var, var, class_method = "quantile",
  nb_class = 9, na.rm = FALSE, replace_min_by_one = FALSE)
}
\arguments{
\item{size}{data.frame containing species and length.}

\item{group_var}{variable to group the data. E.g. species.}

\item{var}{variable containing the values to class.}

\item{class_method}{character percentile or quantile. Default to percentile.}

\item{nb_class}{integer number of size class to create. Default to 9.}

\item{na.rm}{logical. Should the NAs be removed ?}

\item{replace_min_by_one}{logical. In the first class, replace low bound by
1.}
}
\description{
Compute classes
}
\details{
The replace_min_by_one option is to reflect the choice made by
Bonnafé et al. to replace the low bound of the first class (= lowest value)
by one. It leads to slighty different results (env. 0.6% of the links).
}
