% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaweb_build.R
\name{compute_links}
\alias{compute_links}
\title{Compute species links}
\usage{
compute_links(size_class, th_prey_size, piscivory_index, fish_diet_shift,
  resource_diet_shift, species, fish, low_bound, upper_bound,
  fish_resource_method = "overlap", pred_win_method = "midpoint")
}
\arguments{
\item{size_class}{a data.frame generated by compute_classes}

\item{th_prey_size}{data.frame generated by compute_prey_size}

\item{piscivory_index}{data.frame generated by compute_piscivory}

\item{fish_diet_shift}{a data.frame containing species, life stage, lower and
upper bound size for each stage}

\item{resource_diet_shift}{data.frame generated by compute_prey_size}

\item{species}{variable name}

\item{fish}{name of logical piscivory variable. 0 = no piscivory; 1 = piscivory}

\item{low_bound}{variable containing lower limit of the stage}

\item{upper_bound}{variable containing upper limit of the stage. Not used.}
}
\value{
list containing the species-species and the
species-resource interaction matrix
}
\description{
Compute feeding links between species based on their size
}
\details{

}
