% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaweb_build.R
\name{compute_piscivory}
\alias{compute_piscivory}
\title{Compute piscivory link}
\usage{
compute_piscivory(size_class, fish_diet_shift, species, low_bound, upper_bound,
  fish)
}
\arguments{
\item{size_class}{a data.frame generated by compute_classes}

\item{fish_diet_shift}{a data.frame containing species, life stage, lower and
upper bound size for each stage}

\item{species}{variable name}

\item{low_bound}{variable containing lower limit of the stage}

\item{upper_bound}{variable containing upper limit of the stage. Not used.}

\item{fish}{name of logical piscivory variable. 0 = no piscivory; 1 = piscivory}
}
\value{
a data.frame
}
\description{
Determine which size class is piscivorous
}
\details{
If given species class is considered as piscivor if the upper bound
of the class is greater than the mininum size to reach piscivory.
}
