% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaweb_build.R
\name{compute_prey_size}
\alias{compute_prey_size}
\title{Compute prey size}
\usage{
compute_prey_size(class_size, pred_win, species, beta_min, beta_max,
  pred_win_method = "midpoint")
}
\arguments{
\item{class_size}{a data.frame generated by compute_classes}

\item{pred_win}{data.frame containing species and the predation window
parameters (\alpha and \beta).}

\item{species}{}

\item{beta_min}{column name containing beta_min parameters}

\item{beta_max}{column name contains beta_max parameters}

\item{pred_win_method}{character midpoint, overlap or no_overlap. Default to midpoint.}
}
\value{
matrix
}
\description{
Compute prey size for class size species can eat. It is theoretical, indeep,
the information about real diet of the fish is not taken in account here.
}
\details{
For now, the function compute the mininum prey size as follow:
beta_min * min_predator_size. But the general formula is in hte form:
alpha_min + beta_min * min_predator_size. But here, we do not have data about
alpha. We will add it if necessary.
}
