% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_network_build.R
\name{extract_network}
\alias{extract_network}
\title{Extract network}
\usage{
extract_network(data, species, var, metaweb, classes = NULL, link = "flux",
  out_format = "adjacency")
}
\arguments{
\item{data}{a data.frame containing species and size variable.}

\item{species}{species variable name in the dataset.}

\item{var}{variable characterizing the size.}

\item{metaweb}{an object created by the build_metaweb function.}

\item{classes}{data.frame created by the compute_classes function.}
}
\value{
a matrix containing the species class and links between them
}
\description{
Extract local network from a list of species and size
}
