% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_network_build.R
\name{get_size_class}
\alias{get_size_class}
\title{Get size classes for a given species}
\usage{
get_size_class(data, species_name, var, classes)
}
\arguments{
\item{data}{a data.frame containing species and size variable.}

\item{var}{variable characterizing the size.}

\item{classes}{data.frame created by the compute_classes function.}

\item{species}{data.frame containing species and the predation window.}
}
\value{
data.frame containing local network code and interaction matrix.
}
\description{
Get size classes for a given species
}
