% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{get_species}
\alias{get_species}
\title{Get species from nodes}
\usage{
get_species(node_list)
}
\arguments{
\item{list}{}
}
\description{
Get species from nodes
}
\details{
By convention the node naming scheme is set as the following:
 species_code + _ + class_nb. This function is just a wrapper around
 extract_all().
}
