% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot_temporal_network}
\alias{plot_temporal_network}
\title{Temporal graph for a station}
\usage{
plot_temporal_network(net, meta, dead, ...)
}
\arguments{
\item{net}{data.frame contaning year variable and network list}

\item{meta}{an object created by build_metaweb}

\item{dead}{character vector indicating which node are basal}

\item{...}{Other arguments to pass to set_layout_graph}
}
\value{
a tibble containing a list of graph in a net_graph column
}
\description{
Temporal graph for a station
}
\seealso{
set_layout_graph
}
