% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{theme_alain}
\alias{theme_alain}
\title{Personal plot theme}
\usage{
theme_alain()
}
\value{
a ggplot theme
}
\description{
Personal plot theme
}
\details{
This theme was design from the theme developped in the hrbrthemes
package: https://github.com/hrbrmstr/hrbrthemes
}
