% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_network_build.R
\name{build_local_network}
\alias{build_local_network}
\title{Local network building}
\usage{
build_local_network(metaweb, classes, data, species, var, group_var)
}
\arguments{
\item{metaweb}{an object created by the build_metaweb function.}

\item{classes}{data.frame created by the compute_classes function.}

\item{data}{a data.frame containing species and size variable.}

\item{species}{data.frame containing species and the predation window.}

\item{var}{variable characterizing the size.}

\item{group_var}{variable characterizing different communities.}
}
\value{
data.frame containing local network code and interaction matrix.
}
\description{
Local network building
}
\details{

}
